fix: url pattern errors
- switch to restful api
- fix comment editor dirty content
- refactor token middleware
- bugs fix

+++ src/api/article_api.rs
@@ -6,14 +6,16 @@ use log::{debug, info};
 use uuid::Uuid;
 
 use crate::models::articles::{
-    AdminViewRawArticle, DeleteArticlesWithTags, ModifyPublish, QuerySlice, ViewArticle,
+    AdminViewRawArticle, ArticleSummary, ModifyPublish, QuerySlice, ViewArticle,
 };
+use crate::models::token::TokenExtension;
+use crate::util::errors::ErrorCode;
 use crate::{AppState, ArticleList, ArticlesWithTag, EditArticle, NewArticle};
 use actix_web::web;
 
-pub struct AdminArticle;
+pub struct ArticleApi;
 
-impl AdminArticle {
+impl ArticleApi {
     pub fn create_article(
         state: Data<AppState>,
         req: HttpRequest,
@@ -25,9 +27,18 @@ impl AdminArticle {
 
     pub fn delete_article(
         state: Data<AppState>,
+        req: HttpRequest,
         article_id: Path<Uuid>,
     ) -> impl Future<Item = HttpResponse, Error = Error> {
         info!("delete article");
+        // The API is only available for administrator
+        if !TokenExtension::is_admin(&req) {
+            return api_resp_err_with_code!(
+                ErrorCode::PermissionDenied,
+                "Permission denied, this API is for administrator only".to_string()
+            );
+        }
+
         let res = ArticlesWithTag::delete_with_id(state.get_ref(), article_id.into_inner());
         match res {
             Ok(v) => api_resp_data!(v),
@@ -90,10 +101,18 @@ impl AdminArticle {
 
     pub fn update_publish(
         state: Data<AppState>,
-        _req: HttpRequest,
+        req: HttpRequest,
         params: Form<ModifyPublish>,
     ) -> impl Future<Item = HttpResponse, Error = Error> {
         info!("update_publish");
+        // The API is only available for administrator
+        if !TokenExtension::is_admin(&req) {
+            return api_resp_err_with_code!(
+                ErrorCode::PermissionDenied,
+                "Permission denied, this API is for administrator only".to_string()
+            );
+        }
+
         let res = ArticlesWithTag::publish_article(state.get_ref(), &params);
         match res {
             Ok(data) => api_resp_data!(data),
@@ -101,33 +120,46 @@ impl AdminArticle {
         }
     }
 
+    fn view_article(
+        state: Data<AppState>,
+        _req: HttpRequest,
+        params: Path<Uuid>,
+    ) -> impl Future<Item = HttpResponse, Error = Error> {
+        debug!("view_article: {:?}", &params);
+        match ArticlesWithTag::query_article(&state, params.into_inner(), false) {
+            Ok(data) => api_resp_data!(data),
+            Err(err) => api_resp_err!(&*err),
+        }
+    }
+
+    pub fn list_all_article(
+        state: Data<AppState>,
+        _req: HttpRequest,
+        params: Query<QuerySlice>,
+    ) -> impl Future<Item = HttpResponse, Error = Error> {
+        info!("list_all_article");
+        let conn = &state.db.connection();
+        let res = ArticleSummary::list_articles(conn, params.limit, params.offset, false);
+        match res {
+            Ok(data) => api_resp_data!(data),
+            Err(_) => api_resp_err!("list_all_article failed!"),
+        }
+    }
+
     pub fn configure(cfg: &mut web::ServiceConfig) {
         cfg.service(
-            web::resource("article/new").route(web::post().to_async(AdminArticle::create_article)),
+            web::resource("/article")
+                .route(web::post().to_async(Self::create_article)) // create
+                .route(web::put().to_async(Self::edit_article)) // update
+                .route(web::patch().to_async(Self::update_publish)), // publish
         )
         .service(
-            web::resource("article/delete/{article_id}")
-                .route(web::delete().to_async(AdminArticle::delete_article)),
+            web::resource("/article/{article_id}")
+                .route(web::delete().to_async(Self::delete_article)) // delete
+                .route(web::get().to_async(Self::view_article)), // get article
         )
         .service(
-            web::resource("article/admin/view")
-                .route(web::get().to_async(AdminArticle::admin_view_article)),
-        )
-        // user
-        .service(
-            web::resource("article/admin/view_raw")
-                .route(web::post().to_async(AdminArticle::admin_view_raw_article)),
-        )
-        .service(
-            web::resource("article/admin/view_all")
-                .route(web::get().to_async(AdminArticle::admin_list_all_article)),
-        )
-        .service(
-            web::resource("article/edit").route(web::post().to_async(AdminArticle::edit_article)),
-        )
-        .service(
-            web::resource("article/publish")
-                .route(web::get().to_async(AdminArticle::update_publish)),
+            web::resource("/articles").route(web::get().to_async(Self::list_all_article)), // list articles
         );
     }
 }
